Refactor delete function and create a page for it
