+++ src/pages/Homepage.jsx
@@ -1,12 +1,12 @@
 import Restaurants from "./Restaurants";
 
 function Homepage() {
-    return (
-        <>
-            <h1 className='text-2xl'>Homepage</h1>
-            <Restaurants />
-        </>
-    );
+  return (
+    <>
+      <h1 className="text-2xl">Restaurants List</h1>
+      <Restaurants />
+    </>
+  );
 }
 
 export default Homepage;
